BitbucketCommit from BitbucketCommits URI fix.
The URIs for BitbucketCommit from BitbucketCommits#getCommit() and BitbucketCommits#latest() were wrong.

They were created by adding the ref to the base of commits URI.
`/<owner>/<name>/commits` -> `/<owner>/<name>/commits/hash`
But the correct form is `/<owner>/<name>/commit/hash` (commit not commits)

Interesting that whe calling getCommit(), Bitbucket won't complain about it,
and the result was just a json commits with a single commit.
Based on this, the extracted BitbucketCommit json from that result was wrong too. (it shouldn't contain the "values" property).