#1071 `Commands` implemented + Language linked responses.
This PR also introduces linked language responses. This means that a value for a response key could be a link
to a separate file that contains the actual reply. The file, depending on used link protocol, could be fetched locally (`file://`, `classpath:`) or remotely (`http(s)://`).

By doing this, larger and/or marked up replies can be easly maintainted in separate files rather than having them clobbered in the `responses.properites`.