ConditionalResource for JsonStorage.
This PR address the problem of multiple db trips when ConditionalJsonResources
checks for etag and body in JsonStorage. By providing a higher abstraction ConditionalResource, an extension of Resource
and therefore changing the original JsonResources API, all checks are done at once.

The only downside is that Resource API has to be moved to self-api module.